refactor(osd-viewer-demo-v2): create components & pages, Add jotai state mgmt
