+++ apps/viewer-demo-v2/package.json
@@ -4,7 +4,7 @@
   "private": true,
   "dependencies": {
     "@craco/craco": "^6.2.0",
-    "@lunit/osd-react-renderer": "^1.0.1",
+    "@lunit/osd-react-renderer": "^1.1.0",
     "@material-ui/core": "^4.12.3",
     "@material-ui/icons": "^4.11.2",
     "@material-ui/lab": "^4.0.0-alpha.60",
@@ -21,11 +21,13 @@
     "openseadragon": "^2.4.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-router-dom": "^5.3.0",
+    "react-router-dom": "^5.3.4",
     "react-scripts": "4.0.3",
     "styled-components": "^5.3.1",
     "typescript": "^4.1.2",
-    "web-vitals": "^1.0.1"
+    "web-vitals": "^1.0.1",
+    "jotai": "^2.5.0",
+    "react-error-boundary": "4.0.11"
   },
   "scripts": {
     "start": "craco start",
@@ -54,7 +56,7 @@
   "devDependencies": {
     "@types/lodash": "^4.14.178",
     "@types/offscreencanvas": "^2019.7.0",
-    "@types/react-router-dom": "^5.3.1",
+    "@types/react-router-dom": "^5.3.3",
     "@types/styled-components": "^5.1.13"
   }
 }
